RMS-Backend-dev-sheetal/RMS-Backend-dev-sheetal/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
asyncpg==0.30.0
attrs==25.4.0
av==15.1.0
backoff==2.2.1
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.11.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.120.4
flatbuffers==25.9.23
frozenlist==1.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
ics==0.7.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
jiter==0.11.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langfuse==3.9.3
langsmith==0.4.39
livekit==1.0.18
livekit-agents==1.2.18
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-plugins-deepgram==1.2.18
livekit-plugins-elevenlabs==1.2.18
livekit-plugins-openai==1.2.18
livekit-plugins-silero==1.2.18
livekit-protocol==1.1.0
lxml==6.0.2
marshmallow==3.26.1
mcp==1.20.0
motor==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.4
onnxruntime==1.23.1
openai==2.6.1
openai-agents==0.4.2
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pillow==12.0.0
pluggy==1.6.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
PyMsgBox==2.0.1
PyMuPDF==1.26.5
pyperclip==1.11.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
sympy==1.14.0
TatSu==5.13.2
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
types-protobuf==6.32.1.20251105
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
